fix(localized-text): now properly resolving
The liquid filter was not properly resolving text when displayed inside
a ContentPicker but now works fine !!
Preview also works properly now.

There is still an edge case I cannot fix:
When a parent and first widget both have the LocalizedTextPart and the
second widget wants to refer to the paren't LocalizedTextPart.
It will resolve to the first widget LocalizedTextPart.
This is a strange design anyways so it should not be an issue.